service_note/requirements.txt
aiohttp==2.2.0
alembic==0.9.1
amqp==2.3.2
anyjson==0.3.3
appnope==0.1.0
argh==0.26.2
asn1crypto==0.22.0
async-timeout==1.2.1
attrs==17.4.0
Automat==0.6.0
backcall==0.1.0
billiard==3.5.0.3
bleach==2.1.3
celery==4.1.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cocos2d==0.6.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.1
cycler==0.10.0
decorator==4.3.0
dynaconf==1.0.2
entrypoints==0.2.3
enum-compat==0.0.2
eventlet==0.21.0
Fabric==1.13.2
Fabric3==1.13.1.post1
flasgger==0.9.0
Flask==1.0.2
Flask-Admin==1.5.0
Flask-Excel==0.0.5
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gevent==1.2.2
greenlet==0.4.12
grpcio==1.12.1
grpcio-tools==1.12.1
gunicorn==19.7.1
hiredis==0.2.0
html5lib==1.0b10
httplib2==0.10.3
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
json-rpc==1.11.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
kombu==4.2.1
lml==0.0.1
lxml==3.7.3
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
multidict==3.0.0
mysql-connector==2.1.6
nameko==2.9.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.5.0
numpy==1.14.3
olefile==0.44
packaging==16.8
pandocfilters==1.4.2
paramiko==2.1.2
parsel==1.4.0
parso==0.2.1
path.py==11.0.1
pathtools==0.1.2
pbr==4.0.3
pexpect==4.6.0
pickleshare==0.7.4
Pillow==4.1.1
powerline-status==2.5.2
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.17
PyDispatcher==2.0.5
pydocstyle==2.1.1
pyexcel==0.5.0
pyexcel-io==0.4.0
pyexcel-webio==0.0.11
pyexcel-xls==0.4.0
pyflakes==1.6.0
pyglet==1.2.4
Pygments==2.2.0
pylama==7.4.3
pymongo==3.4.0
PyMySQL==0.7.11
pyOpenSSL==17.5.0
pypandoc==1.4
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
pypinyin==0.30.0
pyquery==1.2.17
pytest==3.2.3
python-box==3.2.0
python-dateutil==2.7.2
python-docx==0.8.6
python-dotenv==0.8.2
python-editor==1.0.3
python3-memcached==1.51
pytils==0.3
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
queuelib==1.5.0
redis==2.10.5
reportlab==3.4.0
requests==2.13.0
Scrapy==1.5.0
selectors==0.0.14
selenium==3.3.1
Send2Trash==1.5.0
service-identity==17.0.0
sidecar==0.1.0.dev0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
SQLAlchemy==1.1.6
terminado==0.8.1
testpath==0.3.1
texttable==0.9.0
thrift==0.11.0
toml==0.9.4
tormysql==0.3.4
tornado==4.5.2
torndsession==1.1.5.1
traitlets==4.3.2
tredis==0.7.2
Twisted==18.4.0
user-agent==0.1.9
vine==1.1.4
virtualenv==15.2.0
w3lib==1.19.0
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5.1
weblib==0.1.25
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
WTForms==2.1
xhtml2pdf==0.2b1
xlrd==1.0.0
XlsxWriter==0.9.9
xlwt==1.2.0
xpinyin==0.5.5
yagmail==0.10.212
yarl==0.10.3
you-get==0.4.803
zope.interface==4.5.0
